back/notificacao/requirements.txt
alembic==1.0.5
asn1crypto==0.24.0
astroid==2.1.0
bcrypt==3.1.4
cffi==1.11.5
Click==7.0
colorama==0.4.1
cryptography==2.6.1
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.3.2
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
psycopg2==2.7.6.1
pycparser==2.19
PyJWT==1.7.0
pylint==2.2.2
python-dateutil==2.7.5
python-editor==1.0.3
rope==0.11.0
six==1.11.0
SQLAlchemy==1.2.14
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
unidecode==1.0.23
Pillow==5.4.1
pywebpush=1.9.0